Huge rework of CmdHandler and Bot classes. Change in logical direction of program.
Updated the majority of the way in which the preAfkCheck and startAfkCheck methods are designed to function.

CmdHandler has new constructor that removes the need for constant re-defining of the adminChannel and AfkText Channel. The only information passed now is the startUser.

Embed Message Builder was also created to create "pretty" messages to be sent in text channels. This is due for rework next update.